HW1. corrections4

+++ src/main/java/ru/javawebinar/topjava/model/MealTo.java
@@ -3,13 +3,13 @@ package ru.javawebinar.topjava.model;
 import java.time.LocalDateTime;
 
 public class MealTo {
-    private final long id;
+    private final int id;
     private final LocalDateTime dateTime;
     private final String description;
     private final int calories;
     private final boolean excess;
 
-    public MealTo(long id, LocalDateTime dateTime, String description, int calories, boolean excess) {
+    public MealTo(int id, LocalDateTime dateTime, String description, int calories, boolean excess) {
         this.id = id;
         this.dateTime = dateTime;
         this.description = description;
@@ -17,7 +17,7 @@ public class MealTo {
         this.excess = excess;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
